transcripts_optimized/requirements.txt
backoff
google-generativeai
langchain-text-splitters
pymongo
python-dotenv
requests
tqdm
structlog>=23.1.0
python-json-logger>=2.0.7
